pyproject.toml
[project]
name = "htm-wl-repro"
version = "0.1.0"
description = "Reproducible HTM-WL pipeline to recreate paper results."
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["numpy","pandas","PyYAML"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["htm_wl*"]
